backend/delete-video-function/requirements.txt
google-cloud-functions
functions-framework
google-cloud-storage
